requirements.txt
blinker==1.8.2
cffi==1.17.1
click==8.1.7
cryptocode==0.1
cryptography==43.0.1
et-xmlfile==1.1.0
Flask==3.0.3
Flask-Humanize==0.3.0
Flask-Login==0.6.3
Flask-Moment==1.0.6
Flask-WTF==1.2.1
greenlet==3.0.3
gunicorn==23.0.0
humanize==4.10.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
msoffcrypto-tool==5.4.2
numpy==2.1.1
olefile==0.47
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pycparser==2.22
pycryptodomex==3.20.0
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
typing_extensions==4.12.2
tzdata==2024.1
Werkzeug==3.0.4
WTForms==3.1.2
xlrd==2.0.1
